♻️ Is638/webserver's project and directorv2 plugins refactoring (#3189)
