+++ services/web/server/src/simcore_service_webserver/director_v2_api.py
@@ -4,17 +4,14 @@ PLEASE avoid importing from any other module to access this plugin's functionali
 """
 
 
-from typing import Tuple
-
 from .director_v2_abc import (
     AbstractProjectRunPolicy,
     get_project_run_policy,
     set_project_run_policy,
 )
-from .director_v2_core import (
+from .director_v2_core_computations import (
     ClusterAccessForbidden,
     ClusterNotFoundError,
-    DirectorServiceError,
     create_cluster,
     create_or_update_pipeline,
     delete_cluster,
@@ -22,52 +19,55 @@ from .director_v2_core import (
     get_cluster,
     get_cluster_details,
     get_computation_task,
-    get_service_state,
-    get_services,
-    is_healthy,
     is_pipeline_running,
     list_clusters,
     ping_cluster,
     ping_specific_cluster,
-    projects_networks_update,
-    request_retrieve_dyn_service,
-    restart,
-    retrieve,
-    start_service,
-    stop_service,
-    stop_services,
     update_cluster,
 )
+from .director_v2_core_dynamic_services import (
+    DirectorServiceError,
+    get_dynamic_service_state,
+    get_dynamic_services,
+    request_retrieve_dyn_service,
+    restart_dynamic_service,
+    retrieve,
+    run_dynamic_service,
+    stop_dynamic_service,
+    stop_dynamic_services_in_project,
+    update_dynamic_service_networks_in_project,
+)
+from .director_v2_core_utils import is_healthy
 
 # director-v2 module internal API
-__all__: Tuple[str, ...] = (
+__all__: tuple[str, ...] = (
     "AbstractProjectRunPolicy",
-    "create_or_update_pipeline",
-    "delete_pipeline",
-    "DirectorServiceError",
-    "get_computation_task",
-    "get_project_run_policy",
-    "get_service_state",
-    "get_services",
-    "is_healthy",
-    "is_pipeline_running",
-    "projects_networks_update",
-    "request_retrieve_dyn_service",
-    "restart",
-    "retrieve",
-    "set_project_run_policy",
-    "start_service",
-    "stop_service",
-    "stop_services",
-    "ping_cluster",
-    "ping_specific_cluster",
-    "create_cluster",
-    "list_clusters",
-    "get_cluster",
-    "get_cluster_details",
-    "update_cluster",
-    "delete_cluster",
     "ClusterAccessForbidden",
     "ClusterNotFoundError",
+    "create_cluster",
+    "create_or_update_pipeline",
+    "delete_cluster",
+    "delete_pipeline",
+    "DirectorServiceError",
+    "get_cluster_details",
+    "get_cluster",
+    "get_computation_task",
+    "get_dynamic_service_state",
+    "get_dynamic_services",
+    "get_project_run_policy",
+    "is_healthy",
+    "is_pipeline_running",
+    "list_clusters",
+    "ping_cluster",
+    "ping_specific_cluster",
+    "request_retrieve_dyn_service",
+    "restart_dynamic_service",
+    "retrieve",
+    "run_dynamic_service",
+    "set_project_run_policy",
+    "stop_dynamic_service",
+    "stop_dynamic_services_in_project",
+    "update_cluster",
+    "update_dynamic_service_networks_in_project",
 )
 # nopycln: file
