+++ services/web/server/src/simcore_service_webserver/director_v2_handlers.py
@@ -1,6 +1,6 @@
 import asyncio
 import logging
-from typing import Any, Dict, List, Optional, Set, Tuple
+from typing import Any, Optional
 
 from aiohttp import web
 from models_library.clusters import ClusterID
@@ -13,7 +13,8 @@ from servicelib.json_serialization import json_dumps
 
 from ._meta import api_version_prefix as VTAG
 from .director_v2_abc import get_project_run_policy
-from .director_v2_core import DirectorServiceError, DirectorV2ApiClient
+from .director_v2_core_computations import ComputationsApi
+from .director_v2_exceptions import DirectorServiceError
 from .login.decorators import RQT_USERID_KEY, login_required
 from .security_decorators import permission_required
 from .version_control_db import CommitID
@@ -29,7 +30,7 @@ routes = web.RouteTableDef()
 @permission_required("services.pipeline.*")
 @permission_required("project.read")
 async def start_computation(request: web.Request) -> web.Response:
-    client = DirectorV2ApiClient(request.app)
+    computations = ComputationsApi(request.app)
 
     run_policy = get_project_run_policy(request.app)
     assert run_policy  # nosec
@@ -37,7 +38,7 @@ async def start_computation(request: web.Request) -> web.Response:
     user_id = UserID(request[RQT_USERID_KEY])
     project_id = ProjectID(request.match_info["project_id"])
 
-    subgraph: Set[str] = set()
+    subgraph: set[str] = set()
     force_restart: bool = False  # TODO: deprecate this entry
     cluster_id: NonNegativeInt = 0
 
@@ -55,8 +56,8 @@ async def start_computation(request: web.Request) -> web.Response:
     }
 
     try:
-        running_project_ids: List[ProjectID]
-        project_vc_commits: List[CommitID]
+        running_project_ids: list[ProjectID]
+        project_vc_commits: list[CommitID]
 
         (
             running_project_ids,
@@ -76,15 +77,18 @@ async def start_computation(request: web.Request) -> web.Response:
             else True
         )
 
-        _started_pipelines_ids: Tuple[str] = await asyncio.gather(
-            *[client.start(pid, user_id, **options) for pid in running_project_ids]
+        _started_pipelines_ids: tuple[str] = await asyncio.gather(
+            *[
+                computations.start(pid, user_id, **options)
+                for pid in running_project_ids
+            ]
         )
 
         assert set(_started_pipelines_ids) == set(
             map(str, running_project_ids)
         )  # nosec
 
-        data: Dict[str, Any] = {
+        data: dict[str, Any] = {
             "pipeline_id": project_id,
         }
         # Optional
@@ -110,7 +114,7 @@ async def start_computation(request: web.Request) -> web.Response:
 @permission_required("services.pipeline.*")
 @permission_required("project.read")
 async def stop_computation(request: web.Request) -> web.Response:
-    client = DirectorV2ApiClient(request.app)
+    computations = ComputationsApi(request.app)
     run_policy = get_project_run_policy(request.app)
     assert run_policy  # nosec
 
@@ -118,12 +122,12 @@ async def stop_computation(request: web.Request) -> web.Response:
     project_id = ProjectID(request.match_info["project_id"])
 
     try:
-        project_ids: List[ProjectID] = await run_policy.get_runnable_projects_ids(
+        project_ids: list[ProjectID] = await run_policy.get_runnable_projects_ids(
             request, project_id
         )
         log.debug("Project %s will stop %d variants", project_id, len(project_ids))
 
-        await asyncio.gather(*[client.stop(pid, user_id) for pid in project_ids])
+        await asyncio.gather(*[computations.stop(pid, user_id) for pid in project_ids])
 
         # FIXME: our middleware has this issue
         #
@@ -149,7 +153,7 @@ class ComputationTaskGet(BaseModel):
 @permission_required("services.pipeline.*")
 @permission_required("project.read")
 async def get_computation(request: web.Request) -> web.Response:
-    client = DirectorV2ApiClient(request.app)
+    computations = ComputationsApi(request.app)
     run_policy = get_project_run_policy(request.app)
     assert run_policy  # nosec
 
@@ -157,14 +161,17 @@ async def get_computation(request: web.Request) -> web.Response:
     project_id = ProjectID(request.match_info["project_id"])
 
     try:
-        project_ids: List[ProjectID] = await run_policy.get_runnable_projects_ids(
+        project_ids: list[ProjectID] = await run_policy.get_runnable_projects_ids(
             request, project_id
         )
         log.debug("Project %s will get %d variants", project_id, len(project_ids))
         list_computation_tasks = parse_obj_as(
-            List[ComputationTaskGet],
+            list[ComputationTaskGet],
             await asyncio.gather(
-                *[client.get(project_id=pid, user_id=user_id) for pid in project_ids]
+                *[
+                    computations.get(project_id=pid, user_id=user_id)
+                    for pid in project_ids
+                ]
             ),
         )
         assert len(list_computation_tasks) == len(project_ids)  # nosec
