+++ services/web/server/src/simcore_service_webserver/garbage_collector_core.py
@@ -311,7 +311,9 @@ async def _remove_single_orphaned_service(
             f"{service_host=}",
         )
         try:
-            await director_v2_api.stop_service(app, service_uuid, save_state=False)
+            await director_v2_api.stop_dynamic_service(
+                app, service_uuid, save_state=False
+            )
         except (ServiceNotFoundError, DirectorException) as err:
             logger.warning("Error while stopping service: %s", err)
         return
@@ -363,7 +365,7 @@ async def _remove_single_orphaned_service(
                 save_state = False
             # -------------------------------------------
 
-            await director_v2_api.stop_service(app, service_uuid, save_state)
+            await director_v2_api.stop_dynamic_service(app, service_uuid, save_state)
 
         except (ServiceNotFoundError, DirectorException) as err:
             logger.warning("Error while stopping service: %s", err)
@@ -396,7 +398,7 @@ async def remove_orphaned_services(
 
     running_interactive_services: list[dict[str, Any]] = []
     try:
-        running_interactive_services = await director_v2_api.get_services(app)
+        running_interactive_services = await director_v2_api.get_dynamic_services(app)
     except director_v2_api.DirectorServiceError:
         logger.debug("Could not fetch running_interactive_services")
